requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# YouTube Download
yt-dlp==2023.11.16

# Transcription (Whisper)
faster-whisper==0.10.0
torch==2.1.1
torchaudio==2.1.1

# LLM Integration (Ollama)
ollama==0.1.6
httpx==0.25.2

# Translation (NLLB)
transformers==4.36.0
sentencepiece==0.1.99
sacremoses==0.1.1

# TTS (Silero + Coqui)
silero==0.4.1
TTS==0.22.0

# Video Processing
ffmpeg-python==0.2.0
moviepy==1.0.3
pillow==10.1.0

# Audio Analysis
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.24.3

